Added initial tests

+++ src/Reformulators/RandomWalkerReformulator.hh
@@ -24,8 +24,8 @@ public:
 	SASPlan RebuildSASPlan(SASPlan* reformulatedSAS) override;
 private:
 	std::vector<Path> PerformWalk(PDDLInstance* instance);
-	std::unordered_map<int, EntanglementOccurance> FindEntanglements(std::vector<Path> paths, PDDLInstance* instance);
-	PDDLInstance GenerateMacros(std::unordered_map<int, EntanglementOccurance> candidates, PDDLInstance* instance);
+	std::unordered_map<size_t, EntanglementOccurance> FindEntanglements(std::vector<Path> paths, PDDLInstance* instance);
+	PDDLInstance GenerateMacros(std::unordered_map<size_t, EntanglementOccurance> candidates, PDDLInstance* instance);
 };
 
 #endif
